+++ sdk/communication/communication-email/samples-dev/sendEmailSingleRecipient.ts
@@ -20,8 +20,8 @@ const sendSingleEmail = async (): Promise<void> => {
   const emailClient: EmailClient = new EmailClient(connectionString);
 
   // Create the Email Message to be sent
-  const emailMessage: EmailMessage = {
-    sender: senderAddress,
+  const message: EmailMessage = {
+    senderEmail: senderAddress,
     content: {
       subject: "This is the subject",
       plainText: "This is the body",
@@ -34,7 +34,7 @@ const sendSingleEmail = async (): Promise<void> => {
 
   try {
     // Send the email message
-    const response = await emailClient.send(emailMessage);
+    const response = await emailClient.send(message);
 
     console.log("Message ID: " + response.messageId);
   } catch (error) {