+++ sdk/communication/communication-email/samples-dev/sendEmailWithAttachments.ts
@@ -20,8 +20,8 @@ const sendEmailWithAttachments = async (): Promise<void> => {
   const emailClient: EmailClient = new EmailClient(connectionString);
 
   // Create the Email Message to be sent
-  const emailMessage: EmailMessage = {
-    sender: senderAddress,
+  const message: EmailMessage = {
+    senderEmail: senderAddress,
     content: {
       subject: "This is the subject",
       plainText: "This is the body",
@@ -33,7 +33,7 @@ const sendEmailWithAttachments = async (): Promise<void> => {
     attachments: [
       {
         name: "readme.txt",
-        attachmentType: "txt",
+        type: "txt",
         contentBytesBase64: "ZW1haWwgdGVzdCBhdHRhY2htZW50",
       },
     ],
@@ -41,7 +41,7 @@ const sendEmailWithAttachments = async (): Promise<void> => {
 
   try {
     // Send the email message
-    const response = await emailClient.send(emailMessage);
+    const response = await emailClient.send(message);
 
     console.log("Message ID: " + response.messageId);
   } catch (error) {